Reworked the distance classes to provide a distance Class.
The Distance type encompasses every possible way to express a distance,
so that adding new ways to express a distance does not require every
single type of distance to be repeated. Additionally, made distance type
a parameter, so that adding new types does not result in the same
type of repetition.